feat(health): Update HealthCheckApi class add api to test create jwt token pair, create uuid, validation token and extract userid
