+++ src/main/java/com/swyp3/babpool/global/config/WebMvcInterceptJwtConfig.java
@@ -13,7 +13,7 @@ public class WebMvcInterceptJwtConfig implements WebMvcConfigurer {
     private final JwtTokenInterceptor jwtTokenInterceptor;
     private static final String[] EXCLUDE_PATHS = {
         "/api/user/sign/in", "/api/user/sign/up", "/api/user/sign/out", "/api/user/sign/refresh",
-        "/api/test/connection", "/api/test/jwt/permitted",
+        "/api/test/connection", "/api/test/jwt/permitted", "/api/test/uuid", "/api/test/jwt/tokens",
         "/api/profile/list"
     };
